+++ client/tailwind.config.js
@@ -1,0 +1,13 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+  content: [
+    "./public/index.html",
+    "./src/**/*.{js,jsx,ts,tsx}",
+    "./src/*.js",
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}
+
